app/requirements.txt
psycopg2-binary>=2.8
mysql
pandas
matplotlib
pymssql
